requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.2
chromadb>=0.4.0
uuid>=1.30
tqdm>=4.66.1
torch>=2.0.1
transformers>=4.34.1
langchain-mistralai>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.1.1
numpy>=1.24.3
sentence-transformers>=2.2.2
rouge-score>=0.1.2
evaluate>=0.4.1
scikit-learn>=1.3.1
bert-score>=0.3.13
odfpy>=1.4.1